+++ App.js
@@ -1,10 +1,21 @@
+import { StatusBar } from 'expo-status-bar';
 import React from 'react';
-import BottomNavigator from './src/routes/BottomNavigator'
-import LoginScreen from './src/screens/LoginScreen'
+import { StyleSheet, Text, View } from 'react-native';
 
 export default function App() {
   return (
-    // <BottomNavigator />
-    <LoginScreen />
+    <View style={styles.container}>
+      <Text>Open up App.js to start working on your app!</Text>
+      <StatusBar style="auto" />
+    </View>
   );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
