--- src/constants/colors.js
@@ -1,4 +1,0 @@
-export const colors = {
-  primaryColor: "#FAFAFA",
-  activeTintColor: "#1FCC79"
-}
