--- src/screens/BMRScreen.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
-
-const BMRScreen = () => {
-  return (
-    <View>
-      <Text>BMRScreen</Text>
-    </View>
-  )
-}
-
-export default BMRScreen
-
-const styles = StyleSheet.create({})
