--- src/screens/ExploreScreen.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
-
-const ExploreScreen = () => {
-  return (
-    <View>
-      <Text>ExploreScreen</Text>
-    </View>
-  )
-}
-
-export default ExploreScreen
-
-const styles = StyleSheet.create({})
