--- src/screens/HomeScreen.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
-
-const HomeScreen = () => {
-  return (
-    <View>
-      <Text>Home Screen</Text>
-    </View>
-  )
-}
-
-export default HomeScreen
-
-const styles = StyleSheet.create({})
