--- src/screens/ProfileScreen.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
-
-const ProfileScreen = () => {
-  return (
-    <View>
-      <Text>ProfileScreen</Text>
-    </View>
-  )
-}
-
-export default ProfileScreen
-
-const styles = StyleSheet.create({})
