--- src/screens/ScanScreen.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
-
-const ScanScreen = () => {
-  return (
-    <View>
-      <Text>ScanScreen</Text>
-    </View>
-  )
-}
-
-export default ScanScreen
-
-const styles = StyleSheet.create({})
