--- src/screens/index.js
@@ -1,5 +1,0 @@
-export { default as HomeScreen } from './HomeScreen'
-export { default as ExploreScreen } from './ExploreScreen'
-export { default as ScanScreen } from './ScanScreen'
-export { default as BMRScreen } from './BMRScreen'
-export { default as ProfileScreen } from './ProfileScreen'
